record limit change Int -> Long

+++ src/main/scala/org/embulk/input/DynamoDBUtil.scala
@@ -53,7 +53,7 @@ object DynamoDBUtil {
     var evaluateKey: JMap[String, AttributeValue] = null
 
     val scanLimit: Int = task.getScanLimit
-    val recordLimit: Int = task.getRecordLimit
+    val recordLimit: Long = task.getRecordLimit
     var recordCount: Int = 0
 
     do {
